Don't expose handle in FsEvent::shutdown_observe
There is a risk that a user could pass in an invalid pointer to
`FsEvent::shutdown_observe`, wehich could trigger undefined behavior.
Instead, this makes an API-breaking change to save the runloop in `FsEvent`.
Furthermore, this replaces FsEventRefWrapper with explicit CFRef wrapeprs
that allow these pointers to be passed across threads.

Finally, it updates the tests to wait on the observation threads to make
sure they shut down correctly.